pipeline/pyproject.toml
[project]
name = "fabric-pipeline"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "jsonschema>=4.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "PyYAML>=6.0",
    "Pillow>=10.0",
    "numpy>=1.26",
    "anthropic>=0.40",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
